requirements.txt
python-telegram-bot
sqlite
